Move the tool from CLI to /tools
